livekit-backend/requirements.txt
livekit-agents[openai,groq,noise-cancellation,silero,turn_detector]==1.0.22
python-dotenv>=1.0.0
